Repair pkgconfig configuration
Case ZC-424:  The move of apr to the new location was the easy part.
pkgconfig, on the other hand, is a complicated bit of magic, and the
apr-1-config script needed some massaging, both so our apr could coexist
peacefully with upstream apr, and so the script would spit out the right
configuration options.